fix: move theme out of pages dir
